pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linparakeet"
version = "0.1.0"
description = "Linux voice-to-text tray app powered by NVIDIA Parakeet STT (fully local, fully private)."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Tammas Ryan" }]
keywords = ["speech-to-text", "asr", "parakeet", "nemo", "linux", "tray", "voice", "dictation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "PySide6>=6.4",
    "sounddevice>=0.4.6",
    "numpy>=1.24,<2.0",
    "nemo_toolkit[asr]>=1.23",
    "pyperclip>=1.8",
    "pynput>=1.7",
    "scipy>=1.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/Tam-mas/linparakeet"
Issues = "https://github.com/Tam-mas/linparakeet/issues"

[project.scripts]
linparakeet = "main:main"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
# E/F = pycodestyle + pyflakes, I = isort, B = bugbear.
# pyupgrade (UP) is intentionally off: the codebase consistently uses
# Optional[X] / typing.Union and we don't want a lint-vs-style fight.
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
